+++ gym_mgmt_system/__manifest__.py
@@ -21,7 +21,7 @@
 #############################################################################
 {
     'name': 'GYM Management System',
-    'version': '17.0.1.0.0',
+    'version': '17.0.2.0.0',
     'category': 'Industries',
     'summary': 'GYM Management System For Managing '
                'Membership, Member, Workout Plan, etc',
@@ -41,8 +41,13 @@
         'security/ir.model.access.csv',
         'security/gym_mgmt_system_security.xml',
         'data/ir_sequence_data.xml',
+        'data/gym_membership_cron.xml',
+        'data/trainer_skill_data.xml',
         'wizard/assign_workout.xml',
+        'wizard/gym_management_extension.xml',
+        'views/gym_menu.xml',
         'views/product_template_views.xml',
+        'views/gym_attendance_views.xml',
         'views/res_partner_views.xml',
         'views/exercise_for_views.xml',
         'views/gym_exercise_views.xml',