+++ gym_mgmt_system/models/__init__.py
@@ -33,3 +33,4 @@ from . import trainer_skill
 from . import workout_days
 from . import workout_plan
 from . import workout_plan_option
+from . import gym_attendance